# backend/gateway/openapi.yaml
swagger: '2.0'
info:
  title: Semper API
  description: >
    API Gateway front door for the private Semper Cloud Run service.
    Device-attested routes require headers X-Device-Id, X-Nonce, X-Signature
    (ECDSA-P256 over nonce+METHOD+path+sha256(body)) in addition to the
    end-user Bearer ID token (Authorization or X-Forwarded-Authorization).

    POST /v1/tasks/provision-session is deliberately absent: it is the Cloud
    Tasks callback, delivered straight to the Cloud Run URL with an OIDC token
    for the task invoker service account, and must not be reachable through the
    public front door at all.
  version: '1.0'
schemes:
  - https
produces:
  - application/json
# All requests are forwarded to the private Cloud Run service, invoked with the
# gateway's backend-auth service account. The incoming path is appended to the
# address, so the backend sees the same path.
#
# Template — replace before `gcloud api-gateway api-configs create`:
#   CLOUD_RUN_URL=$(gcloud run services describe indic-api --region REGION \
#                     --format='value(status.url)')
#   # FIREBASE_PROJECT_ID must be the Auth project (e.g. indicvision-dic-app-auth),
#   # which can differ from GOOGLE_CLOUD_PROJECT (e.g. indicvision-dic-app).
#   sed -e "s|__CLOUD_RUN_URL__|${CLOUD_RUN_URL}|g" \
#       -e "s|__FIREBASE_PROJECT_ID__|${FIREBASE_PROJECT_ID}|g" \
#       openapi.yaml > openapi.generated.yaml
# Never commit a live Cloud Run hostname or project id here.
...
#
...
securityDefinitions:
  firebase:
    authorizationUrl: ''
    flow: implicit
    type: oauth2
    x-google-issuer: https://securetoken.google.com/__FIREBASE_PROJECT_ID__
    x-google-jwks_uri: https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com
    x-google-audiences: __FIREBASE_PROJECT_ID__
# Applies to every operation; the health probes opt out with `security: []`.
security:
  - firebase: []
# Per-consumer quotas (ESPv2 / Service Management). Deploy with the gateway
# OpenAPI config so these metrics are registered; in-process token buckets in
# app/rate_limit.py are belt-and-braces per Cloud Run instance.
...
paths:
...
  /v1/licenses/activate:
    post:
      operationId: activateLicense
      description: >
        Redeem a license key. For an individual key, locked to this account's
        email and device. For an institution key (kind=institution), the
        caller's verified email domain must match the license's domainLock;
        a seat is created (or re-bound to this device) on first activation,
        subject to maxSeats. Idempotent re-entry of the same key from the
        already-bound uid/device succeeds without consuming a new seat.
      responses:
        '200': { description: OK }
        '403': { description: "Lock mismatch, revoked, disabled seat, or domain mismatch (license_email_mismatch, license_device_mismatch, license_revoked, license_seat_disabled)" }
        '404': { description: Unknown key }
        '409': { description: "Already redeemed by another account, or institution seats exhausted (license_already_redeemed, license_seats_exhausted)" }
        '429': { description: Rate limited }
...